refactor: ItemList 분리하기
