+++ src/components/index.ts
@@ -1,3 +1,4 @@
 import { Header } from "./Header";
+import { ItemList } from "./ItemList";
 
-export { Header };
+export { Header, ItemList };
