+++ src/types/index.ts
@@ -3,3 +3,10 @@ export interface User {
   name: string;
   email: string;
 }
+
+export interface Item {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
